requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.0
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.4
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
Django==5.0.3
django-allauth==0.61.1
django-cleanup==8.1.0
django-htmx==1.17.3
einops==0.7.0
filelock==3.13.3
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.34
langchain-text-splitters==0.0.1
langsmith==0.1.33
llama-index==0.10.23
llama-index-agent-openai==0.1.7
llama-index-cli==0.1.11
llama-index-core==0.10.23.post1
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.12
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.4
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.12
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.0
llamaapi==0.1.36
llamaindex-py-client==0.1.13
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.14.3
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
psutil==5.9.8
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pypdf==4.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.18.1
sentence-transformers==2.6.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
sqlparse==0.4.4
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.39.1
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
